chore(struct): for linux.
