+++ server/lib/main.cpp
@@ -1,29 +1,30 @@
 #include <iostream>
 #include "sorry/router.hpp"
 #include "sorry/service.hpp"
-
 using namespace web::http;
 
 int main()
 {
-    experimental::listener::http_listener listener{router::getEndPoint()};
 
-    router r{listener};
-    service s{r};
-    try
-    {
-        std::cout << "sorry server start" << std::endl;
-        listener.open().wait();
+	std::cout << "sorry progress start" << std::endl;
+	experimental::listener::http_listener listener{ router::getEndPoint() };
 
-        std::string line;
-        std::getline(std::cin, line);
+	router r{ listener };
+	service s{ r };
+	try
+	{
+		std::cout << "sorry server start" << std::endl;
+		listener.open().wait();
 
-        std::cout << "sorry server close" << std::endl;
-        listener.close().wait();
-    }
-    catch (std::exception &e)
-    {
-        std::cout << "Error occurred sending response: %s\n"
-                  << e.what();
-    }
+		std::string line;
+		std::getline(std::cin, line);
+
+		std::cout << "sorry server close" << std::endl;
+		listener.close().wait();
+	}
+	catch (std::exception &e)
+	{
+		std::cout << "Error occurred sending response: %s\n"
+			<< e.what();
+	}
 }
