--- server/sorry/log/file.hpp
@@ -1,11 +1,0 @@
-#ifndef LOGFILE_H
-#define LOGFILE_H
-#include <iostream>
-#include <fstream>
-
-class file
-{
-  public:
-	  static void write();
-};
-#endif //LOGFILE_H
