--- server/sorry/log/log.hpp
@@ -1,9 +1,0 @@
-
-#include "./mongo.hpp"
-
-namespace log
-{
-class trunk
-{
-};
-} // namespace log
